pyproject.toml
[tool.poetry]
name = "repl-nix-workspace"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.2"
uvicorn = "^0.30.6"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
